requirements.txt
cffi
adal
azure
pytest
vcrpy
